Fix most of the new vs2017 induced warnings.
Refs: Phab:D3096 Phab:D3103 #5862
Reviewed by: @wraitii
Comments by: @Angen
Differential Revision: https://code.wildfiregames.com/D3126

git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@24268 3db68df2-c116-0410-a063-a993310a9797

+++ source/graphics/ParticleEmitter.cpp
@@ -1,4 +1,4 @@
-/* Copyright (C) 2015 Wildfire Games.
+/* Copyright (C) 2020 Wildfire Games.
  * This file is part of 0 A.D.
  *
  * 0 A.D. is free software: you can redistribute it and/or modify
@@ -69,7 +69,7 @@ CParticleEmitter::CParticleEmitter(const CParticleEmitterTypePtr& type) :
 	m_IndexArray.SetNumVertices(m_Type->m_MaxParticles * 6);
 	m_IndexArray.Layout();
 	VertexArrayIterator<u16> index = m_IndexArray.GetIterator();
-	for (size_t i = 0; i < m_Type->m_MaxParticles; ++i)
+	for (u16 i = 0; i < m_Type->m_MaxParticles; ++i)
 	{
 		*index++ = i*4 + 0;
 		*index++ = i*4 + 1;
